Add files via upload
ATM:
This program assignment will involve the idea of modern-day banking using an ATM.
The server-side software maintains a bank account and a user
accesses the account via client software.
There is two sides to this program: a server and a client. The server maintains the balance
and does all updates (withdrawals and deposits) and also responds to queries on the amount of
money in the account. The user issues commands via the client to check the available balance,
withdraw money from the account, and deposit money into the account. It does none of the
calculations but only sends requests to the server that executes the commands and returns the
answer to the client. The server should initialize the balance in the bank account to $10,000.
2. The user can be able to do the following activities from the client software:
a. Deposit money into the account on the server.
b. Withdraw money from the account on the server.
c. Check the balance of the amount in the account on the server.